move productsParams methods to Product.helpers file
