+++ src/services/products.ts
@@ -1,7 +1,5 @@
 import { useQuery } from 'react-query';
-import { RouteComponentProps } from 'react-router-dom';
 import { api } from 'utils/apiClient';
-import { queryStringToObject, objectToQueryString } from 'utils/queryString';
 import { ProductsParams, ProductsResponseData } from './products.types';
 
 const fetchProducts = async (params: ProductsParams = {}): Promise<ProductsResponseData> => {
@@ -21,25 +19,4 @@ export const useProducts = (params: ProductsParams = {}, options: Options = {}) 
     keepPreviousData: true,
     ...options,
   });
-};
-
-// URL query params handling
-export const productsParamsFromURL = (queryParamsFromUrl: string): ProductsParams => {
-  const { page, active, promo, search } = queryStringToObject(queryParamsFromUrl);
-
-  return {
-    ...(search && { search }),
-    ...(active === 'true' && { active: true }),
-    ...(promo === 'true' && { promo: true }),
-    ...(page && !Number.isNaN(parseInt(page, 10)) ? { page: parseInt(page, 10) } : { page: 1 }),
-  };
-};
-
-type History = RouteComponentProps['history'];
-
-export const productsParamsToURL = (params: ProductsParams, history: History, pathname: string) => {
-  const paramsToURL = { ...params };
-  if (paramsToURL.page === 1) delete paramsToURL.page;
-
-  history.replace({ pathname, search: objectToQueryString(paramsToURL) });
 };
